creating exam part and question part

+++ src/main/java/com/quiz/quiz_exam/repository/StudentAnswerRepository.java
@@ -3,5 +3,5 @@ package com.quiz.quiz_exam.repository;
 import com.quiz.quiz_exam.entity.StudentAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentAnswerRepository extends JpaRepository<StudentAnswer,Integer> {
+public interface StudentAnswerRepository extends JpaRepository<StudentAnswer,Long> {
 }
